show deadlines in their own section in summary_cran_results
the previous version of the code incorrectly assumed that a package with a deadline had failing CRAN checks. It turns out there are packages that have a deadline but pass all CRAN checks. This change takes this into account, by not flagging these packages as "all clear", and instead provide separate output for deadlines instead of adding the deadline to the output of failing CRAN checks